1.) both templating modules now utilize the same response function 2.) cleaned up dead code in tests
