+++ src/extensions/templating/mustache.jl
@@ -4,23 +4,6 @@ using .Mustache
 
 export mustache
 
-
-"""
-    response(content::String, status=200, headers=[]) :: HTTP.Response
-
-Convert a template string `content` into a valid HTTP Response object.
-The content type header is automatically generated based on the content's mimetype
-- `content`: The string content to be included in the HTTP response body.
-- `status`: The HTTP status code (default is 200).
-- `headers`: Additional HTTP headers to include (default is an empty array).
-
-Returns an `HTTP.Response` object with the specified content, status, and headers.
-"""
-function response(content::String, status=200, headers=[]; detect=true) :: HTTP.Response
-    response_headers = detect ? ["Content-Type" => HTTP.sniff(content)] : []
-    return HTTP.Response(status, [response_headers; headers;], content)
-end
-
 """
     mustache(template::String; kwargs...)
